Move all Svelte components to /components
