+++ frontend/src/lib/components/Card.svelte
