+++ frontend/src/lib/components/Footer.svelte
