+++ frontend/src/lib/components/NavBar.svelte
